avoid return statement in outermost code block
